move ScalarMaybeUndef into the miri engine
